requirements.txt
ipykernel
streamlit 
langchain 
langchain_ollama
selenium
beautifulsoup4
lxml 
html5lib
python-dotenv
langchain-groq
webdriver-manager